requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == 'Windows'
pymupdf>=1.19.0  # For PDF text extraction
pytesseract>=0.3.10  # For OCR
Pillow>=10.0.0  # For image processing
python-docx>=1.0.0  # For DOCX files
python-pptx>=0.6.21  # For PPT/PPTX files
langchain>=0.0.200  # For AI agents
langchain-community>=0.0.10  # For community integrations
langchain-core>=0.1.0  # Core LangChain functionality
langchain-text-splitters>=0.0.1  # For text processing
langchain-openai>=0.0.1  # For OpenAI integration (if needed)
google-generativeai>=0.3.0  # Google's Generative AI SDK
langchain-google-genai>=0.0.5  # LangChain integration for Google Generative AI
python-dotenv>=1.0.0  # For loading environment variables
pydantic>=2.0.0  # For data validation
tqdm>=4.65.0  # For progress bars
python-multipart>=0.0.6  # For file uploads
fastapi>=0.95.0  # For API endpoints (if needed)
uvicorn>=0.21.0  # ASGI server (if using FastAPI)
loguru>=0.7.0  # For logging
python-slugify>=8.0.0  # For generating URL-friendly names
pyyaml>=6.0.0  # For YAML configuration files
numpy>=1.24.0  # For numerical operations
pandas>=2.0.0  # For data manipulation (if needed)

# Development Dependencies
pytest>=7.3.1
pytest-cov>=4.0.0
black>=23.3.0
isort>=5.12.0
mypy>=1.2.0
flake8>=6.0.0
pre-commit>=3.2.2
